requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm[asyncpg]==0.20.0
aerich==0.7.2
celery==5.3.4
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-decouple==3.8
pydantic[email]==2.5.0
asyncpg==0.29.0